requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
idna==3.11
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
optuna==4.7.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
psutil==7.2.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sklearn-compat==0.1.5
SQLAlchemy==2.0.45
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.4.0
torch==2.9.1
torch-geometric==2.7.0
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
